https://github.com/opencart/opencart/issues/5048

+++ upload/install/opencart.sql
@@ -941,7 +941,7 @@ CREATE TABLE `oc_currency` (
   `symbol_left` varchar(12) NOT NULL,
   `symbol_right` varchar(12) NOT NULL,
   `decimal_place` char(1) NOT NULL,
-  `value` float(15,8) NOT NULL,
+  `value` double(15,8) NOT NULL,
   `status` tinyint(1) NOT NULL,
   `date_modified` datetime NOT NULL,
   PRIMARY KEY (`currency_id`)
